+++ packages/api-reference/src/components/Content/Content.vue
@@ -1,14 +1,18 @@
 <script setup lang="ts">
+import { ScalarErrorBoundary } from '@scalar/components'
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { ApiReferenceConfiguration } from '@scalar/types'
 import type { WorkspaceStore } from '@scalar/workspace-store/client'
 import { computed } from 'vue'
 
-import { Introduction } from '@/components/Content/Introduction'
 import { Models } from '@/components/Content/Models'
 import { SectionFlare } from '@/components/SectionFlare'
 import { useConfig } from '@/hooks/useConfig'
+import { ScalarAuthSelector } from '@/v2/blocks/scalar-auth-selector'
+import { ScalarClientSelector } from '@/v2/blocks/scalar-client-selector'
+import { ScalarDocumentInfo } from '@/v2/blocks/scalar-document-info'
 import { generateClientOptions } from '@/v2/blocks/scalar-request-example-block/helpers/generate-client-options'
+import { ScalarServerSelector } from '@/v2/blocks/scalar-server-selector'
 
 import { TraversedEntryContainer } from './Operations'
 
@@ -34,7 +38,7 @@ const clientOptions = computed(() =>
     <slot name="start" />
 
     <!-- Introduction -->
-    <Introduction
+    <ScalarDocumentInfo
       :document="store.workspace.activeDocument"
       :clientOptions
       :defaultClient="store.workspace['x-scalar-default-client']"
@@ -42,7 +46,22 @@ const clientOptions = computed(() =>
       <template #empty-state>
         <slot name="empty-state" />
       </template>
-    </Introduction>
+      <template #selectors>
+        <ScalarErrorBoundary>
+          <ScalarServerSelector :config="config" />
+        </ScalarErrorBoundary>
+        <ScalarErrorBoundary>
+          <ScalarAuthSelector :config="config" />
+        </ScalarErrorBoundary>
+        <ScalarErrorBoundary>
+          <ScalarClientSelector
+            :config="config"
+            :clientOptions="clientOptions"
+            :selectedClient="store.workspace['x-scalar-default-client']"
+            :document="store.workspace.activeDocument" />
+        </ScalarErrorBoundary>
+      </template>
+    </ScalarDocumentInfo>
 
     <!-- Loop on traversed entries -->
     <TraversedEntryContainer
