--- packages/api-reference/src/components/Content/Introduction/index.ts
@@ -1,1 +1,0 @@
-export { default as Introduction } from './Introduction.vue'
