+++ packages/api-reference/src/components/IntroductionCard/index.ts
@@ -1,0 +1,2 @@
+export { default as IntroductionCard } from './IntroductionCard.vue'
+export { default as IntroductionCardItem } from './IntroductionCardItem.vue'
