+++ packages/api-reference/src/v2/blocks/scalar-auth-selector/index.ts
@@ -1,0 +1,1 @@
+export { default as ScalarAuthSelector } from './ScalarAuthSelector.vue'
