+++ packages/api-reference/src/v2/blocks/scalar-document-info/InfoDescription.vue
