+++ packages/api-reference/src/v2/blocks/scalar-document-info/OpenApiVersion.vue
