+++ packages/api-reference/src/v2/blocks/scalar-document-info/index.ts
@@ -1,0 +1,1 @@
+export { default as ScalarDocumentInfo } from './ScalarDocumentInfo.vue'
