+++ packages/api-reference/src/v2/blocks/scalar-server-selector/index.ts
@@ -1,0 +1,1 @@
+export { default as ScalarServerSelector } from './ScalarServerSelector.vue'
